Move supabase client into utils
