+++ src/components/channel-forms.js
@@ -1,5 +1,5 @@
 import {useState} from 'react'
-import {supabase} from '../supabaseClient'
+import {supabase} from '../utils/supabaseClient'
 
 export function CreateForm({onCreate}) {
 	const [loading, setLoading] = useState(false)
